LPS-89635 Move to new package
